refactor: update dependency files with their respective correct names
